support double underscore
