+++ tests/endtoend/snake_case_functions/valid_stein/function_app.py
@@ -54,3 +54,41 @@ def sandwich(_req_: func.HttpRequest)\
         -> func.HttpResponse:
     name = _req_.params.get('name')
     return func.HttpResponse(f"Hello, {name}.")
+
+
+@app.route(route="double_underscore", trigger_arg_name="req__snake")
+def double_underscore(req__snake: func.HttpRequest) -> func.HttpResponse:
+    name = req__snake.params.get('name')
+    return func.HttpResponse(f"Hello, {name}.")
+
+
+@app.route(route="double_underscore_prefix", trigger_arg_name="__req")
+def classic_double_underscore(__req: func.HttpRequest) -> func.HttpResponse:
+    name = __req.params.get('name')
+    return func.HttpResponse(f"Hello, {name}.")
+
+
+def double_underscore_suffix(req__: func.HttpRequest) -> func.HttpResponse:
+    name = req__.params.get('name')
+    return func.HttpResponse(f"Hello, {name}.")
+
+
+@app.route(route="jsut_double_underscore", trigger_arg_name="__")
+def jsut_double_underscore(__: func.HttpRequest) -> func.HttpResponse:
+    name = __.params.get('name')
+    return func.HttpResponse(f"Hello, {name}.")
+
+
+@app.route(route="python_main_keyword", trigger_arg_name="__main__")
+def python_main_keyword(__main__: func.HttpRequest) -> func.HttpResponse:
+    name = __main__.params.get('name')
+    return func.HttpResponse(f"Hello, {name}.")
+
+
+@app.route(route="ultimate_combo2",
+           trigger_arg_name="__9req__snake__sna_ke________snake__sn0ke_")
+def ultimate_combo2(
+        __9req__snake__sna_ke________snake__sn0ke_: func.HttpRequest)\
+        -> func.HttpResponse:
+    name = __9req__snake__sna_ke________snake__sn0ke_.params.get('name')
+    return func.HttpResponse(f"Hello, {name}.")
