+++ tests/endtoend/test_snake_case_functions.py
@@ -99,180 +99,68 @@ class TestValidSnakeCaseFunctions(testutils.WebHostTestCase):
             b'Hello, query.'
         )
 
-
-class TestValidSnakeCaseFunctionsStein(TestValidSnakeCaseFunctions):
-
-    @classmethod
-    def get_script_dir(cls):
-        return testutils.E2E_TESTS_FOLDER / 'snake_case_functions' / \
-                                            'valid_stein'
-
-
-class TestInvalidSnakeCaseFunctions(testutils.WebHostTestCase):
-    def setUp(self):
-        self._patch_environ = patch.dict('os.environ', os.environ.copy())
-        self._patch_environ.start()
-        super().setUp()
-
-    def tearDown(self):
-        super().tearDown()
-        self._patch_environ.stop()
-
-    @classmethod
-    def get_script_dir(cls):
-        return testutils.E2E_TESTS_FOLDER / 'snake_case_functions'
-
     @testutils.retryable_test(3, 5)
     def test_double_underscore(self):
         r = self.webhost.request('GET', 'double_underscore',
                                  params={'name': 'query'},
                                  timeout=REQUEST_TIMEOUT_SEC)
-        self.assertFalse(r.ok)
+        self.assertTrue(r.ok)
+        self.assertEqual(
+            r.content,
+            b'Hello, query.'
+        )
 
     @testutils.retryable_test(3, 5)
     def test_double_underscore_prefix(self):
         r = self.webhost.request('GET', 'double_underscore_prefix',
                                  params={'name': 'query'},
                                  timeout=REQUEST_TIMEOUT_SEC)
-        self.assertFalse(r.ok)
+        self.assertTrue(r.ok)
+        self.assertEqual(
+            r.content,
+            b'Hello, query.'
+        )
 
     @testutils.retryable_test(3, 5)
     def test_double_underscore_suffix(self):
         r = self.webhost.request('GET', 'double_underscore_suffix',
                                  params={'name': 'query'},
                                  timeout=REQUEST_TIMEOUT_SEC)
-        self.assertFalse(r.ok)
+        self.assertTrue(r.ok)
+        self.assertEqual(
+            r.content,
+            b'Hello, query.'
+        )
 
     @testutils.retryable_test(3, 5)
     def test_just_double_underscore(self):
         r = self.webhost.request('GET', 'just_double_underscore',
                                  params={'name': 'query'},
                                  timeout=REQUEST_TIMEOUT_SEC)
-        self.assertFalse(r.ok)
+        self.assertTrue(r.ok)
+        self.assertEqual(
+            r.content,
+            b'Hello, query.'
+        )
 
     @testutils.retryable_test(3, 5)
     def test_python_main_keyword(self):
         r = self.webhost.request('GET', 'python_main_keyword',
                                  params={'name': 'query'},
                                  timeout=REQUEST_TIMEOUT_SEC)
-        self.assertFalse(r.ok)
-
-
-class TestInvalidSnakeCaseFunctionsV21(testutils.WebHostTestCase):
-    def setUp(self):
-        self._patch_environ = patch.dict('os.environ', os.environ.copy())
-        self._patch_environ.start()
-        super().setUp()
-
-    def tearDown(self):
-        super().tearDown()
-        self._patch_environ.stop()
-
-    @classmethod
-    def get_script_dir(cls):
-        return testutils.E2E_TESTS_FOLDER / 'snake_case_functions' / \
-                                            'invalid_stein' / \
-                                            'double_underscore'
+        self.assertTrue(r.ok)
+        self.assertEqual(
+            r.content,
+            b'Hello, query.'
+        )
 
     @testutils.retryable_test(3, 5)
-    def test_double_underscore(self):
-        r = self.webhost.request('GET', 'double_underscore',
+    def test_ultimate_combo2(self):
+        r = self.webhost.request('GET', 'ultimate_combo2',
                                  params={'name': 'query'},
                                  timeout=REQUEST_TIMEOUT_SEC)
-        self.assertFalse(r.ok)
-
-
-class TestInvalidSnakeCaseFunctionsV22(testutils.WebHostTestCase):
-    def setUp(self):
-        self._patch_environ = patch.dict('os.environ', os.environ.copy())
-        self._patch_environ.start()
-        super().setUp()
-
-    def tearDown(self):
-        super().tearDown()
-        self._patch_environ.stop()
-
-    @classmethod
-    def get_script_dir(cls):
-        return testutils.E2E_TESTS_FOLDER / 'snake_case_functions' / \
-            'invalid_stein' / \
-            'double_underscore_prefix'
-
-    @testutils.retryable_test(3, 5)
-    def test_double_underscore_prefix(self):
-        r = self.webhost.request('GET', 'double_underscore_prefix',
-                                 params={'name': 'query'},
-                                 timeout=REQUEST_TIMEOUT_SEC)
-        self.assertFalse(r.ok)
-
-
-class TestInvalidSnakeCaseFunctionsV23(testutils.WebHostTestCase):
-    def setUp(self):
-        self._patch_environ = patch.dict('os.environ', os.environ.copy())
-        self._patch_environ.start()
-        super().setUp()
-
-    def tearDown(self):
-        super().tearDown()
-        self._patch_environ.stop()
-
-    @classmethod
-    def get_script_dir(cls):
-        return testutils.E2E_TESTS_FOLDER / 'snake_case_functions' / \
-            'invalid_stein' / \
-            'double_underscore_suffix'
-
-    @testutils.retryable_test(3, 5)
-    def test_double_underscore_suffix(self):
-        r = self.webhost.request('GET', 'double_underscore_suffix',
-                                 params={'name': 'query'},
-                                 timeout=REQUEST_TIMEOUT_SEC)
-        self.assertFalse(r.ok)
-
-
-class TestInvalidSnakeCaseFunctionsV24(testutils.WebHostTestCase):
-    def setUp(self):
-        self._patch_environ = patch.dict('os.environ', os.environ.copy())
-        self._patch_environ.start()
-        super().setUp()
-
-    def tearDown(self):
-        super().tearDown()
-        self._patch_environ.stop()
-
-    @classmethod
-    def get_script_dir(cls):
-        return testutils.E2E_TESTS_FOLDER / 'snake_case_functions' / \
-            'invalid_stein' / \
-            'just_double_underscore'
-
-    @testutils.retryable_test(3, 5)
-    def test_just_double_underscore(self):
-        r = self.webhost.request('GET', 'just_double_underscore',
-                                 params={'name': 'query'},
-                                 timeout=REQUEST_TIMEOUT_SEC)
-        self.assertFalse(r.ok)
-
-
-class TestInvalidSnakeCaseFunctionsV25(testutils.WebHostTestCase):
-    def setUp(self):
-        self._patch_environ = patch.dict('os.environ', os.environ.copy())
-        self._patch_environ.start()
-        super().setUp()
-
-    def tearDown(self):
-        super().tearDown()
-        self._patch_environ.stop()
-
-    @classmethod
-    def get_script_dir(cls):
-        return testutils.E2E_TESTS_FOLDER / 'snake_case_functions' / \
-            'invalid_stein' / \
-            'python_main_keyword'
-
-    @testutils.retryable_test(3, 5)
-    def test_python_main_keyword(self):
-        r = self.webhost.request('GET', 'python_main_keyword',
-                                 params={'name': 'query'},
-                                 timeout=REQUEST_TIMEOUT_SEC)
-        self.assertFalse(r.ok)
+        self.assertTrue(r.ok)
+        self.assertEqual(
+            r.content,
+            b'Hello, query.'
+        )
